Attempt to fix Bool Recording
I get an error when trying to ploayback a bool on the record node. It's trying to lerp the value, but bools can't use the lerp function.
Trying to figure out a fix via skipping the smoothing on bools.
Current issue is when playing back, it starts activated until the end of the time it was first hit. After that it seems fine.
Unsure if it's an issue with the output I'm passing or an unrelated bug somewhere else.