Attempt to fix Bool Recording
I get an error when trying to playback a bool on the record node. It's trying to lerp the value, but bools can't use the lerp function.
Added a quick fix via skipping the smoothing when detecting a bool.

Planning to add a better way to not check the variable type every frame. I imagine using a flag that's set once when it's created then checking that bool would be better than a near constant type check. Too tired to do it right now though.